+++ src/main/scala/snakes/controller/controllerComponent/ICommand.scala
@@ -1,0 +1,6 @@
+package snakes.controller.controllerComponent
+
+trait ICommand {
+  def doStep(): Unit
+  def undoStep(): Unit
+}
