+++ src/main/scala/snakes/model/boardComponent/IBoard.scala
@@ -1,0 +1,7 @@
+package snakes.model.boardComponent
+
+trait IBoard {
+  def getSize: Int
+  def getSnakes: Map[Int, Int]
+  def getLadders: Map[Int, Int]
+}
