+++ src/main/scala/snakes/model/boardComponent/IBoardFactory.scala
@@ -1,0 +1,5 @@
+package snakes.model.boardComponent
+
+trait IBoardFactory {
+  def createBoard(size: Int, snakeStrategy: IFeatureGenerationStrategy, ladderStrategy: IFeatureGenerationStrategy): IBoard
+}
