+++ src/main/scala/snakes/model/boardComponent/IFeatureGenerationStrategy.scala
@@ -1,0 +1,5 @@
+package snakes.model.boardComponent
+
+trait IFeatureGenerationStrategy {
+  def createFeatures(size: Int): Map[Int, Int]
+}
