+++ src/test/scala/ControllerSpec.scala
@@ -2,8 +2,10 @@ package snakes.controller
 
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers
-
+import snakes.controller.controllerComponent.Controller
 import snakes.model.*
+import snakes.model.gameComponent.aGame
+import snakes.model.playerComponent.Player
 
 class ControllerSpec extends AnyWordSpec with Matchers {
